backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart>=0.0.18
google-generativeai==0.3.2
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0
pyttsx3==2.90
gtts==2.5.1
python-dotenv==1.0.0
flask-cors==4.0.0

